Refactor: seperate checklist service
